added javadoc and redoing pom.xml file
