+++ src/main/java/io/numis/service/package-info.java
@@ -1,0 +1,14 @@
+/**
+ * Provides the class necessary to instantiate
+ * NumisService provider for numis.io.
+ * <p>
+ * numis.io consists of two entities:
+ * The application and backend service.
+ * The application is a native mobile app
+ * for both iOS and Android.
+ * 
+ * TODO: Update doc.
+ * 
+ * @since 0.0.1
+ */
+package io.numis.service;
